pyproject.toml
[project]
name = "gen-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "hf-xet>=1.2.0",
    "huggingface-hub[hf-xet]>=0.36.0",
    "langchain>=1.2.7",
    "langchain-anthropic>=1.3.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.53.1",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.57.6",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
